docstrings, rename viewer_param.py to viewer_elements.py
