+++ nerfstudio/viewer/server/viewer_state.py
@@ -39,7 +39,7 @@ from nerfstudio.viewer.server.render_state_machine import (
     RenderStateMachine,
 )
 from nerfstudio.viewer.server.subprocess import get_free_port
-from nerfstudio.viewer.server.viewer_param import ViewerElement
+from nerfstudio.viewer.server.viewer_elements import ViewerElement
 from nerfstudio.viewer.viser import ViserServer
 from nerfstudio.viewer.viser._messages import (
     CameraMessage,
